refactor: router core
