+++ deploy/kubernetes/aibrix/semantic-router/config.yaml
@@ -115,6 +115,12 @@ categories:
       - model: vllm-llama3-8b-instruct
         score: 0.7
         use_reasoning: false
+  - name: thinking
+    system_prompt: "You are a thinking expert, should think multiple steps before answering. Please answer the question step by step."
+    model_scores:
+      - model: vllm-llama3-8b-instruct
+        score: 0.7
+        use_reasoning: true
 
 default_model: vllm-llama3-8b-instruct
 
@@ -175,6 +181,11 @@ classifier:
     use_cpu: true
     pii_mapping_path: "models/pii_classifier_modernbert-base_presidio_token_model/pii_type_mapping.json"
 
+keyword_rules:
+  - category: "thinking"
+    operator: "OR"
+    keywords: ["urgent", "immediate", "asap", "think", "careful"]
+    case_sensitive: false
 
 # Router Configuration for Dual-Path Selection
 router:
